+++ CHANGELOG.md
@@ -12,6 +12,7 @@ experienced users of SimFin.
 -   Added 'full' variants for datasets: income, balance, cashflow.
     The 'full' variant has more columns than the 'free' variant.
 -   Added 'shareprices-latest' dataset variant.
+-   Added support for dataset markets e.g. `sf.load_income(variant='annual', market='us')`
 
 
 ## Version 0.1.0 (2019-09-26)